move models out of __init__ to comply with Django 1.9
